fix(antigravity): keep client profile constants out of webpack client bundle
Move profile types and normalization into shared constants so the dashboard
page no longer imports open-sse services that depend on node:crypto.
