+++ src/shared/constants/antigravityClientProfile.ts
@@ -1,15 +1,8 @@
-import {
-  ANTIGRAVITY_CLIENT_PROFILE_VALUES,
-  DEFAULT_ANTIGRAVITY_CLIENT_PROFILE,
-  normalizeAntigravityClientProfile,
-  type AntigravityClientProfile,
-} from "@omniroute/open-sse/services/antigravityClientProfile.ts";
+export const ANTIGRAVITY_CLIENT_PROFILE_VALUES = ["ide", "harness"] as const;
 
-export {
-  ANTIGRAVITY_CLIENT_PROFILE_VALUES,
-  DEFAULT_ANTIGRAVITY_CLIENT_PROFILE,
-  type AntigravityClientProfile,
-};
+export type AntigravityClientProfile = (typeof ANTIGRAVITY_CLIENT_PROFILE_VALUES)[number];
+
+export const DEFAULT_ANTIGRAVITY_CLIENT_PROFILE: AntigravityClientProfile = "ide";
 
 export type AntigravityClientProfileSetting = AntigravityClientProfile;
 
@@ -20,5 +13,17 @@ export const ANTIGRAVITY_CLIENT_PROFILE_OPTIONS: Array<{
   { value: "ide", labelKey: "antigravityClientProfileIde" },
   { value: "harness", labelKey: "antigravityClientProfileHarness" },
 ];
+
+function toNonEmptyString(value: unknown): string | null {
+  return typeof value === "string" && value.trim().length > 0 ? value.trim() : null;
+}
+
+export function normalizeAntigravityClientProfile(value: unknown): AntigravityClientProfile {
+  const normalized = toNonEmptyString(value)?.toLowerCase();
+  if (normalized === "harness" || normalized === "cli" || normalized === "sdk") {
+    return "harness";
+  }
+  return DEFAULT_ANTIGRAVITY_CLIENT_PROFILE;
+}
 
 export const normalizeAntigravityClientProfileSetting = normalizeAntigravityClientProfile;
